Enum chip_8::instructions::Instruction

pub enum Instruction {

    Sys(Addr),
    Clear,
    Return,
    Jump(Addr),
    Call(Addr),
    SkipEqualOperand(Vx, Byte),
    SkipNotEqualOperand(Vx, Byte),
    SkipEqual(Vx, Vy),
    LoadOperand(Vx, Byte),
    AddOperand(Vx, Byte),
    Load(Vx, Vy),
    Or(Vx, Vy),
    And(Vx, Vy),
    XOr(Vx, Vy),
    Add(Vx, Vy),
    Sub(Vx, Vy),
    ShiftRight(Vx, Vy),
    SubNegated(Vx, Vy),
    ShiftLeft(Vx, Vy),
    SkipNotEqual(Vx, Vy),
    LoadI(Addr),
    LongJump(Addr),
    Random(Vx, Byte),
    Draw(Vx, Vy, Nibble),
    SkipKeyPressed(Vx),
    SkipKeyNotPressed(Vx),
    LoadRegisterDelayTimer(Vx),
    LoadKey(Vx),
    LoadDelayTimerRegister(Vx),
    LoadSoundTimerRegister(Vx),
    AddI(Vx),
    LoadSprite(Vx),
    LoadBinaryCodedDecimal(Vx),
    LoadMemoryRegisters(Vx),
    LoadRegistersMemory(Vx),

}

Byte code instruction

Variants

Sys(Addr)

Tuple Fields

0: Addr

Jumps to machine routine at Addr

0nnn - SYS addr

Clear

Clears the display

00E0 - CLS

Return

Returns from a subroutine

00EE - RET

Jump(Addr)

Tuple Fields

0: Addr

Jumps to Addr

1nnn - JP addr

Call(Addr)

Tuple Fields

0: Addr

Calls subroutine at Addr

2nnn - CALL addr

SkipEqualOperand(Vx, Byte)

Tuple Fields

0: Vx

1: Byte

Skips next instruction if Vx equals byte

3xkk - SE Vx, byte

SkipNotEqualOperand(Vx, Byte)

Tuple Fields

0: Vx

1: Byte

Skips next instruction if Vx is not equal to byte

4xkk - SNE Vx, byte

SkipEqual(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Skips next instruction if Vy is equal to Vy

5xy0 - SE Vx, Vy

LoadOperand(Vx, Byte)

Tuple Fields

0: Vx

1: Byte

Loads byte into Vx

6xkk - LD Vx, byte

AddOperand(Vx, Byte)

Tuple Fields

0: Vx

1: Byte

Adds byte to Vx, then stores it in Vx

7xkk - ADD Vx, byte

Load(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Loads Vy into Vx

8xy0 - LD Vx, Vy

Or(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Sets Vx to Vx OR Vy

8xy1 - OR Vx, Vy

And(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Sets Vx to Vx AND Vy

8xy2 - AND Vx, Vy

XOr(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Sets Vx to Vx XOR Vy

8xy3 - XOR Vx, Vy

Add(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Sets Vx to Vx + Vy, VF to carry

8xy4 - ADD Vx, Vy

Sub(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Sets Vx to Vx - Vy, VF to not borrow

8xy5 - SUB Vx, Vy

ShiftRight(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Sets Vx to Vy SHR 1

8xy6 - SHR Vx {, Vy}

SubNegated(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Sets Vx to Vy - Vx, VF to not borrow

8xy7 - SUBN Vx, Vy

ShiftLeft(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Sets Vx to Vy SHL 1

8xyE - SHL Vx {, Vy}

SkipNotEqual(Vx, Vy)

Tuple Fields

0: Vx

1: Vy

Skips next instruction if Vx is not equal to Vy

9xy0 - SNE Vx, Vy

LoadI(Addr)

Tuple Fields

0: Addr

Loads Addr into register I

Annn - LD I, addr

LongJump(Addr)

Tuple Fields

0: Addr

Jumps to Addr + V0

Bnnn - JP V0, addr

Random(Vx, Byte)

Tuple Fields

0: Vx

1: Byte

Sets Vx to random number AND kk

Cxkk - RND Vx, byte

Draw(Vx, Vy, Nibble)

Tuple Fields

0: Vx

1: Vy

2: Nibble

Read n bytes of memory from address I, draw it at Vx and Vy screen coordinates and set VF for erased pixels

Dxyn - DRW Vx, Vy, nibble

SkipKeyPressed(Vx)

Tuple Fields

0: Vx

Skip next instruction if key Vx is pressed

Ex9E - SKP Vx

SkipKeyNotPressed(Vx)

Tuple Fields

0: Vx

Skip next instruction if key Vx is not pressed

ExA1 - SKNP Vx

LoadRegisterDelayTimer(Vx)

Tuple Fields

0: Vx

Set Vx to delay timer value

Fx07 - LD Vx, DT

LoadKey(Vx)

Tuple Fields

0: Vx

Wait for key press and store it in Vx

Fx0A - LD Vx, K

LoadDelayTimerRegister(Vx)

Tuple Fields

0: Vx

Set delay timer to Vx

Fx15 - LD DT, Vx

LoadSoundTimerRegister(Vx)

Tuple Fields

0: Vx

Set sound timer to Vx

Fx18 - LD ST, Vx

AddI(Vx)

Tuple Fields

0: Vx

Add Vx to I

Fx1E - ADD I, Vx

LoadSprite(Vx)

Tuple Fields

0: Vx

Set I to the address of the sprite Vx

Fx29 - LD F, Vx

LoadBinaryCodedDecimal(Vx)

Tuple Fields

0: Vx

Store binary-coded decimal (BCD) at I, I+1 and I+2

Fx33 - LD B, Vx

LoadMemoryRegisters(Vx)

Tuple Fields

0: Vx

Store registers V0..Vx in memory at I

Fx55 - LD [I], Vx

LoadRegistersMemory(Vx)

Tuple Fields

0: Vx

Read registers V0..Vx from memory at I

Fx65 - LD Vx, [I]

Implementations

impl Instruction

pub fn decode(bits: u16) -> Result<Self>

Decodes raw bits into a valid Instruction

Errors

Will return Chip8Error::UnknownInstruction if given bits don’t match any known instruction.

pub fn encode(&self) -> u16

Encodes a valid Instruction into raw bits